core: use current date/time with microseconds as line identifier
The identifier of a line is now a "long long" variable with the current time
(microseconds precision), instead of a counter in the buffer wrapping at
INT_MAX.

It is guaranteed to be unique in the buffer and strictly increasing, and it
persists after /upgrade: when the upgrade file was written by an older
version, the identifier is rebuilt from the date of print saved in the file.

On buffers with free content, the identifier is still the line number ("y").

Buffer property "next_line_id" is replaced by "lines_last_id_assigned", read
with buffer_get_longlong instead of buffer_get_integer.

Breaking changes in API:

- hdata "line_data": type of variable "id" is now "long long"
- infolist "buffer_lines": type of variable "id" is now "long long"
- infolist "buffer": variable "next_line_id" is replaced by
  "lines_last_id_assigned" ("long long")
- function line_search_by_id: type of parameter "id" is now "long long"
- relay "api" protocol: version bumped to 0.7.0
- relay "weechat" protocol: type of line "id" is now "lon" instead of "int"

+++ src/plugins/perl/weechat-perl-api.c
@@ -4023,7 +4023,7 @@ API_FUNC(line_search_by_id)
     result = API_PTR2STR(
         weechat_line_search_by_id (
             API_STR2PTR(SvPV_nolen (ST (0))), /* buffer */
-            SvIV (ST (1)))); /* id */
+            (long long)(SvIV (ST (1))))); /* id */
 
     API_RETURN_STRING(result);
 }
